Move constant PDB tests to PDBConstantTests.cs
