+++ src/Compilers/CSharp/Test/Emit/PDB/PDBConstantTests.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System.Globalization;
+using System.IO;
 using Microsoft.CodeAnalysis.CSharp.Test.Utilities;
 using Microsoft.CodeAnalysis.Emit;
 using Microsoft.CodeAnalysis.Test.Utilities;
@@ -14,7 +15,244 @@ namespace Microsoft.CodeAnalysis.CSharp.UnitTests.PDB
         private CultureInfo _testCulture = new CultureInfo("en-US");
 
         [Fact]
-        public void TestSimpleLocalConstants()
+        public void StringsWithSurrogateChar()
+        {
+            var source = @"
+using System;
+public class T
+{
+    public static void Main()
+    {
+        const string HighSurrogateCharacter = ""\uD800"";
+        const string LowSurrogateCharacter = ""\uDC00"";
+        const string MatchedSurrogateCharacters = ""\uD800\uDC00"";
+    }
+}";
+
+            var c = CreateCompilationWithMscorlibAndSystemCore(source, options: TestOptions.DebugDll);
+
+            // Note:  U+FFFD is the Unicode 'replacement character' point and is used to replace an incoming character
+            //        whose value is unknown or unrepresentable in Unicode.  This is what our pdb writer does with
+            //        unpaired surrogates.
+            c.VerifyPdb(@"
+<symbols>
+  <methods>
+    <method containingType=""T"" name=""Main"">
+      <customDebugInfo>
+        <using>
+          <namespace usingCount=""1"" />
+        </using>
+      </customDebugInfo>
+      <sequencePoints>
+        <entry offset=""0x0"" startLine=""6"" startColumn=""5"" endLine=""6"" endColumn=""6"" />
+        <entry offset=""0x1"" startLine=""10"" startColumn=""5"" endLine=""10"" endColumn=""6"" />
+      </sequencePoints>
+      <scope startOffset=""0x0"" endOffset=""0x2"">
+        <namespace name=""System"" />
+        <constant name=""HighSurrogateCharacter"" value=""\uFFFD"" type=""String"" />
+        <constant name=""LowSurrogateCharacter"" value=""\uFFFD"" type=""String"" />
+        <constant name=""MatchedSurrogateCharacters"" value=""\uD800\uDC00"" type=""String"" />
+      </scope>
+    </method>
+  </methods>
+</symbols>", format: DebugInformationFormat.Pdb);
+
+            c.VerifyPdb(@"
+<symbols>
+  <methods>
+    <method containingType=""T"" name=""Main"">
+      <sequencePoints>
+        <entry offset=""0x0"" startLine=""6"" startColumn=""5"" endLine=""6"" endColumn=""6"" />
+        <entry offset=""0x1"" startLine=""10"" startColumn=""5"" endLine=""10"" endColumn=""6"" />
+      </sequencePoints>
+      <scope startOffset=""0x0"" endOffset=""0x2"">
+        <constant name=""HighSurrogateCharacter"" value=""\uD800"" type=""String"" />
+        <constant name=""LowSurrogateCharacter"" value=""\uDC00"" type=""String"" />
+        <constant name=""MatchedSurrogateCharacters"" value=""\uD800\uDC00"" type=""String"" />
+      </scope>
+    </method>
+  </methods>
+</symbols>", format: DebugInformationFormat.PortablePdb);
+        }
+
+        [Fact, WorkItem(546862, "http://vstfdevdiv:8080/DevDiv2/DevDiv/_workitems/edit/546862")]
+        public void InvalidUnicodeString()
+        {
+            var source = @"
+using System;
+public class T
+{
+    public static void Main()
+    {
+        const string invalidUnicodeString = ""\uD800\0\uDC00"";
+    }
+}";
+
+            var c = CreateCompilationWithMscorlibAndSystemCore(source, options: TestOptions.DebugDll);
+
+            // Note:  U+FFFD is the Unicode 'replacement character' point and is used to replace an incoming character
+            //        whose value is unknown or unrepresentable in Unicode.  This is what our pdb writer does with
+            //        unpaired surrogates.
+            c.VerifyPdb(@"
+<symbols>
+  <methods>
+    <method containingType=""T"" name=""Main"">
+      <customDebugInfo>
+        <using>
+          <namespace usingCount=""1"" />
+        </using>
+      </customDebugInfo>
+      <sequencePoints>
+        <entry offset=""0x0"" startLine=""6"" startColumn=""5"" endLine=""6"" endColumn=""6"" />
+        <entry offset=""0x1"" startLine=""8"" startColumn=""5"" endLine=""8"" endColumn=""6"" />
+      </sequencePoints>
+      <scope startOffset=""0x0"" endOffset=""0x2"">
+        <namespace name=""System"" />
+        <constant name=""invalidUnicodeString"" value=""\uFFFD\u0000\uFFFD"" type=""String"" />
+      </scope>
+    </method>
+  </methods>
+</symbols>", format: DebugInformationFormat.Pdb);
+
+            c.VerifyPdb(@"
+<symbols>
+  <methods>
+    <method containingType=""T"" name=""Main"">
+      <sequencePoints>
+        <entry offset=""0x0"" startLine=""6"" startColumn=""5"" endLine=""6"" endColumn=""6"" />
+        <entry offset=""0x1"" startLine=""8"" startColumn=""5"" endLine=""8"" endColumn=""6"" />
+      </sequencePoints>
+      <scope startOffset=""0x0"" endOffset=""0x2"">
+        <constant name=""invalidUnicodeString"" value=""\uD800\u0000\uDC00"" type=""String"" />
+      </scope>
+    </method>
+  </methods>
+</symbols>", format: DebugInformationFormat.PortablePdb);
+        }
+
+        [Fact]
+        public void AllTypes()
+        {
+            var source = @"
+using System;
+using System.Collections.Generic;
+
+class X {}
+
+public class C<S>
+{
+    enum EnumI1 : sbyte  { A }
+    enum EnumU1 : byte   { A }
+    enum EnumI2 : short  { A }
+    enum EnumU2 : ushort { A }
+    enum EnumI4 : int    { A }
+    enum EnumU4 : uint   { A }
+    enum EnumI8 : long   { A }
+    enum EnumU8 : ulong  { A }
+
+    public static void F<T>()
+    {
+        const bool B = false;
+        const char C = '\0';
+        const sbyte I1 = 0;
+        const byte U1 = 0;
+        const short I2 = 0;
+        const ushort U2 = 0;
+        const int I4 = 0;
+        const uint U4 = 0;
+        const long I8 = 0;
+        const ulong U8 = 0;
+        const float R4 = 0;
+        const double R8 = 0;
+
+        const C<int>.EnumI1 EI1 = 0;
+        const C<int>.EnumU1 EU1 = 0;
+        const C<int>.EnumI2 EI2 = 0;
+        const C<int>.EnumU2 EU2 = 0;
+        const C<int>.EnumI4 EI4 = 0;
+        const C<int>.EnumU4 EU4 = 0;
+        const C<int>.EnumI8 EI8 = 0;
+        const C<int>.EnumU8 EU8 = 0;
+
+        const string StrWithNul = ""\0"";
+        const string EmptyStr = """";
+        const string NullStr = null;
+        const object NullObject = null;
+        const dynamic NullDynamic = null;
+        const X NullTypeDef = null;
+        const Action NullTypeRef = null;
+        const Func<Dictionary<int, C<int>>, dynamic, T, List<S>> NullTypeSpec = null;
+        const object[] Array1 = null;
+        const object[,] Array2 = null;
+        const object[][] Array3 = null;
+
+        const decimal D = 0M;
+        // DateTime const not expressible in C#
+    }
+}";
+
+            var c = CreateCompilationWithMscorlibAndSystemCore(source, options: TestOptions.DebugDll);
+
+            c.VerifyPdb("C`1.F", @"
+<symbols>
+  <methods>
+    <method containingType=""C`1"" name=""F"">
+      <customDebugInfo>
+        <using>
+          <namespace usingCount=""2"" />
+        </using>
+        <dynamicLocals>
+          <bucket flagCount=""1"" flags=""1"" slotId=""0"" localName=""NullDynamic"" />
+          <bucket flagCount=""9"" flags=""000001000"" slotId=""0"" localName=""NullTypeSpec"" />
+        </dynamicLocals>
+      </customDebugInfo>
+      <sequencePoints>
+        <entry offset=""0x0"" startLine=""19"" startColumn=""5"" endLine=""19"" endColumn=""6"" />
+        <entry offset=""0x1"" startLine=""56"" startColumn=""5"" endLine=""56"" endColumn=""6"" />
+      </sequencePoints>
+      <scope startOffset=""0x0"" endOffset=""0x2"">
+        <namespace name=""System"" />
+        <namespace name=""System.Collections.Generic"" />
+        <constant name=""B"" value=""0"" type=""Boolean"" />
+        <constant name=""C"" value=""0"" type=""Char"" />
+        <constant name=""I1"" value=""0"" type=""SByte"" />
+        <constant name=""U1"" value=""0"" type=""Byte"" />
+        <constant name=""I2"" value=""0"" type=""Int16"" />
+        <constant name=""U2"" value=""0"" type=""UInt16"" />
+        <constant name=""I4"" value=""0"" type=""Int32"" />
+        <constant name=""U4"" value=""0"" type=""UInt32"" />
+        <constant name=""I8"" value=""0"" type=""Int64"" />
+        <constant name=""U8"" value=""0"" type=""UInt64"" />
+        <constant name=""R4"" value=""0"" type=""Single"" />
+        <constant name=""R8"" value=""0"" type=""Double"" />
+        <constant name=""EI1"" value=""0"" signature=""EnumI1{Int32}"" />
+        <constant name=""EU1"" value=""0"" signature=""EnumU1{Int32}"" />
+        <constant name=""EI2"" value=""0"" signature=""EnumI2{Int32}"" />
+        <constant name=""EU2"" value=""0"" signature=""EnumU2{Int32}"" />
+        <constant name=""EI4"" value=""0"" signature=""EnumI4{Int32}"" />
+        <constant name=""EU4"" value=""0"" signature=""EnumU4{Int32}"" />
+        <constant name=""EI8"" value=""0"" signature=""EnumI8{Int32}"" />
+        <constant name=""EU8"" value=""0"" signature=""EnumU8{Int32}"" />
+        <constant name=""StrWithNul"" value=""\u0000"" type=""String"" />
+        <constant name=""EmptyStr"" value="""" type=""String"" />
+        <constant name=""NullStr"" value=""null"" type=""String"" />
+        <constant name=""NullObject"" value=""null"" type=""Object"" />
+        <constant name=""NullDynamic"" value=""null"" type=""Object"" />
+        <constant name=""NullTypeDef"" value=""null"" signature=""X"" />
+        <constant name=""NullTypeRef"" value=""null"" signature=""System.Action"" />
+        <constant name=""NullTypeSpec"" value=""null"" signature=""System.Func`4{System.Collections.Generic.Dictionary`2{Int32, C`1{Int32}}, Object, !!0, System.Collections.Generic.List`1{!0}}"" />
+        <constant name=""Array1"" value=""null"" signature=""Object[]"" />
+        <constant name=""Array2"" value=""null"" signature=""Object[,,]"" />
+        <constant name=""Array3"" value=""null"" signature=""Object[][]"" />
+        <constant name=""D"" value=""0"" type=""Decimal"" />
+      </scope>
+    </method>
+  </methods>
+</symbols>");
+        }
+
+        [Fact]
+        public void SimpleLocalConstant()
         {
             var text = @"
 class C
@@ -55,7 +293,7 @@ class C
         }
 
         [Fact]
-        public void TestLambdaLocalConstants()
+        public void LambdaLocalConstants()
         {
             var text = @"
 using System;
@@ -122,7 +360,7 @@ class C
 
         [WorkItem(543342, "http://vstfdevdiv:8080/DevDiv2/DevDiv/_workitems/edit/543342")]
         [Fact]
-        public void TestIteratorLocalConstants()
+        public void IteratorLocalConstants()
         {
             var source = @"
 using System.Collections.Generic;
@@ -202,7 +440,7 @@ class C
         }
 
         [Fact]
-        public void TestLocalConstantsTypes()
+        public void LocalConstantsTypes()
         {
             var text = @"
 class C
@@ -241,6 +479,53 @@ class C
   </methods>
 </symbols>");
             }
+        }
+
+        [Fact]
+        public void WRN_PDBConstantStringValueTooLong()
+        {
+            var longStringValue = new string('a', 2049);
+            var source = @"
+using System;
+
+class C
+{
+    static void Main()
+    {
+        const string foo = """ + longStringValue + @""";
+        Console.Write(foo);
+    }
+}
+";
+
+            var compilation = CreateCompilationWithMscorlib(source, options: TestOptions.DebugExe);
+
+            var exebits = new MemoryStream();
+            var pdbbits = new MemoryStream();
+            var result = compilation.Emit(exebits, pdbbits);
+            result.Diagnostics.Verify();
+
+            /*
+             * old behavior. This new warning was abandoned
+            result.Diagnostics.Verify(// warning CS7063: Constant string value of 'foo' is too long to be used in a PDB file. Only the debug experience may be affected.
+                                      Diagnostic(ErrorCode.WRN_PDBConstantStringValueTooLong).WithArguments("foo", longStringValue.Substring(0, 20) + "..."));
+
+            //make sure that this warning is suppressable
+            compilation = CreateCompilationWithMscorlib(text, compOptions: Options.Exe.WithDebugInformationKind(Common.DebugInformationKind.Full).WithOptimizations(false).
+                WithSpecificDiagnosticOptions(new Dictionary<int, ReportWarning>(){ {(int)ErrorCode.WRN_PDBConstantStringValueTooLong, ReportWarning.Suppress} }));
+
+            result = compilation.Emit(exebits, null, "DontCare", pdbbits, null);
+            result.Diagnostics.Verify();
+
+            //make sure that this warning can be turned into an error.
+            compilation = CreateCompilationWithMscorlib(text, compOptions: Options.Exe.WithDebugInformationKind(Common.DebugInformationKind.Full).WithOptimizations(false).
+                WithSpecificDiagnosticOptions(new Dictionary<int, ReportWarning>() { { (int)ErrorCode.WRN_PDBConstantStringValueTooLong, ReportWarning.Error } }));
+
+            result = compilation.Emit(exebits, null, "DontCare", pdbbits, null);
+            Assert.False(result.Success);
+            result.Diagnostics.Verify(
+                                      Diagnostic(ErrorCode.WRN_PDBConstantStringValueTooLong).WithArguments("foo", longStringValue.Substring(0, 20) + "...").WithWarningAsError(true));
+             * */
         }
 
         [Fact]
@@ -327,7 +612,7 @@ this is a string constant that is too long to fit into the PDB"";
         }
 
         [Fact, WorkItem(178988, "http://vstfdevdiv:8080/DevDiv2/DevDiv/_workitems/edit/178988")]
-        public void TestStringWithNulCharacter_MaxSupportedLength()
+        public void StringWithNulCharacter_MaxSupportedLength()
         {
             const int length = 2031;
             string str = new string('x', 9) + "\0" + new string('x', length - 10);
@@ -380,7 +665,7 @@ class C
         }
 
         [Fact, WorkItem(178988, "http://vstfdevdiv:8080/DevDiv2/DevDiv/_workitems/edit/178988")]
-        public void TestStringWithNulCharacter_OverSupportedLength()
+        public void StringWithNulCharacter_OverSupportedLength()
         {
             const int length = 2032;
             string str = new string('x', 9) + "\0" + new string('x', length - 10);
@@ -430,7 +715,7 @@ class C
         }
 
         [Fact]
-        public void TestDecimalLocalConstants()
+        public void DecimalLocalConstants()
         {
             var text = @"
 class C
